Access control implemented

+++ Assignment1/src/main/java/com/mcino/assignment1/controller/StudentModuleController.java
@@ -2,6 +2,8 @@ package com.mcino.assignment1.controller;
 
 import com.mcino.assignment1.Utils.StudentGrade;
 import com.mcino.assignment1.Utils.StudentGradesForm;
+import com.mcino.assignment1.exception.ModuleNotFoundException;
+import com.mcino.assignment1.model.Module;
 import com.mcino.assignment1.model.StudentModule;
 import com.mcino.assignment1.repository.StudentModuleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +13,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 import java.util.ArrayList;
 import java.util.List;
@@ -23,7 +26,10 @@ public class StudentModuleController {
 
     @RequestMapping("/students/{student_id}/drop/{module_id}")
     public String unEnrollFromModule(@PathVariable(value = "module_id") Long moduleId,
-                                                @PathVariable(value = "student_id") Long studentId) {
+                                     @PathVariable(value = "student_id") Long studentId, HttpSession session) {
+        if(!session.getAttribute("role").equals("student") || ((long) session.getAttribute("id")) != studentId){
+            return "redirect:/error";
+        }
         StudentModule studentModule = studentModuleRepository.findByStudentIdAndModuleId(studentId, moduleId);
         studentModuleRepository.delete(studentModule);
         return "redirect:/my_profile";
@@ -31,7 +37,18 @@ public class StudentModuleController {
 
     @PostMapping("/grades/{module_id}/update/")
     public String updateGrade(@PathVariable(value = "module_id") Long moduleId,
-                              @Valid @ModelAttribute("studentModules")StudentGradesForm studentGradesForm) {
+                              @Valid @ModelAttribute("studentModules")StudentGradesForm studentGradesForm, HttpSession session) {
+
+        if(!session.getAttribute("role").equals("staff")){
+            return "redirect:/error";
+        }
+
+        Module module = studentModuleRepository.findByModuleId(moduleId).get(0).getModule();
+
+        if( ((long) session.getAttribute("id")) != module.getCoordinatorId()){
+            return "redirect:/error";
+        }
+
         List<StudentModule> studentModules = new ArrayList<>();
         StudentModule sm;
         for(StudentGrade sg : studentGradesForm.getStudentGrades()) {
